moving ReceiptItem class into model_objects.py
